push

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -8,5 +8,5 @@
  */
 int print_last_digit(int b)
 {
-	return (abs(b)%10);
+	return (abs((int) b) % 10);
 }
